Adapt filter chips to match Material 3 design
