+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed:
+- Adapt filter chips to match Material 3 design
+
+## [0.8.365] - 2023-06-02
 ### Fixed:
 - Display of Whisper model sizes
